+++ app/controllers/tasks_controller.rb
@@ -15,7 +15,7 @@ class TasksController < ApplicationController
   # GET /projects/1/tasks
   def index
     @order = scrub_order(Task, params[:order], "tasks.created_at desc")
-    @tasks = viewable_tasks.search(params[:search]).order(@order)
+    @tasks = viewable_tasks.search_any_order(params[:search]).order(@order)
                            .page(params[:page]).per(40)
   end