+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 ### Bug Fix
 - Admins can now search for users by last name on users index
 - AM fields are no longer cutoff in time entry fields in certain browsers
+- New sheets on events no longer disassociate from the event when the form
+  language is changed
 
 ## 64.1.1 (June 27, 2018)