+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 65.0.0
+## 65.0.0 (September 10, 2018)
 
 ### Enhancements
 - **General Changes**